Delta Temperature

Rearing temperature in the North and South Delta

delta_temps

Format

dataframe with 504 rows and 3 variables

date

calendar date, monthly value 1980-1999

watershed

CVPIA watershed name

monthly_mean_temp_c

modeled monthly mean temperature in °C

Source

Data Wrangling and Additional Temperature Modeling: Sadie Gill sgill@flowwest.com

Details

Water temperature in the Delta was modeled as a function of air temperature.

North Delta was modeled using a sample of recent measured water temperature at CDEC station EMM and NOAA CDO air temperature data from Antioch.

South Delta was modeled using a sample of recent measured water temperature at CDEC station MTB and NOAA CDO air temperature data Stockton Airport.

Air Temperature Data Source:

NOAA Climate Data Online (CDO) accessed using the rnoaa R package developed by rOpenSci.

Stream Water Temperature Data Sources:

DWR California Data Exchange Center (CDEC) accessed using the CDECRetrieve R packaged developed by FlowWest.