Monthly Mean Water Temperature

A dataset containing the monthly mean water temperature in celsius for each CVPIA watershed 1980-2000.

juv_temp

Format

dataframe with 7,440 rows and 3 variables

date

calendar date, monthly value 1980-2000

watershed

CVPIA watershed name

monthly_mean_temp_c

modeled monthly mean temperature in °C

Source

HEC-5Q Model Output and Double Regression Temperature Modeling: Michael Wright mwright@usbr.gov

Data Wrangling and Additional Temperature Modeling: Sadie Gill sgill@flowwest.com

Details

The following four methods were used to estimate the monthly mean water temperature for the watersheds during 1980-2000:

HEC-5Q
Temperature assignments were determined by Mike Wright using the input files for the three HEC-5Q runs, i.e. AR_5QCS.dat and AR_5Q-CL.OUT.

Empirical Data
Only the Lower Sacramento had sufficient measured water temperature during the period of the CVPIA salmon life cycle model. The few missing values were imputed using forecast::na.interp.

Double Regression Modeling
Data from many temperature gauges was collected from several CVPIA watersheds. In these watersheds, a double regression technique was employed to predict water temperature based on equilibrium temperature and the river mile of interest. Equilibrium temperature is the temperature a pool of water would approach if left in contact with the air for an infinite time, generally equivalent to the air temperature. The results were applied to both watersheds of origin and as surrogate data for ungauged watersheds.

Modeling Details

Additional Temperature Modeling
For streams without major dams, water temperature is highly correlated with air temperature. For each watershed not included in the HEC-5Q model that had partial water temperature data, a linear model was fitted to estimate water temperature as a function of air temperature.

Generally, the air temperature record spans both the period of the CVPIA salmon life cycle model and the complete period of record of available water temperature data. In the cases where there were missing air temperature values between 1980-2000, we imputed the missing values using forecast::na.interp in order to have a complete air temperature dataset for water temperature estimation.

Air Temperature Data Source:

NOAA Climate Data Online (CDO) accessed using the rnoaa R package developed by rOpenSci.

Stream Water Temperature Data Sources:

DWR California Data Exchange Center (CDEC) accessed using the CDECRetrieve R packaged developed by FlowWest.

USGS National Water Information System (NWIS) accessed using the dataRetrieval R package developed by USGS.

Watershed Modeling Details

Upper Sacramento River HEC-5Q model output at COTTONWOOD CR in SSJB_SAC_Reference_062315/SAC/SAC_CL_TEMP.DSS

Antelope Creek Regression method was fitted to data from in-river temperature gauges

Battle Creek HEC-5Q model output at BATTLE CR in SSJB_SAC_Reference_062315/SAC/SAC_CL_TEMP.DSS

Bear Creek Regression method was fitted to data from in-river temperature gauges

Big Chico Creek Estimated mean monthly water temperature from a linear model fitted with water temperature data from CDEC Gage ID: BIC and air temperature data from NOAA CDO Station Id: USC00046685

Butte Creek Estimated mean monthly water temperature from a linear model fitted with water temperature data from CDEC Gage ID: BCD and air temperature data from NOAA CDO Station Id: USC00046685

Clear Creek HEC-5Q model output at IGO in SSJB_SAC_Reference_062315/SAC/SAC_CL_TEMP.DSS

Cottonwood Creek HEC-5Q model output at COTTONWOOD CR in SSJB_SAC_Reference_062315/SAC/SAC_CL_TEMP.DSS

Cow Creek HEC-5Q model output at COW CR in SSJB_SAC_Reference_062315/SAC/SAC_CL_TEMP.DSS

Deer Creek Estimated mean monthly water temperature from a linear model fitted with water temperature data from USGS Gage ID: 11383500 and air temperature data from NOAA CDO Station Id: USC00041715

Elder Creek Regression method was fitted to data from Antelope River temperature gauges

Mill Creek Estimated mean monthly water temperature from a linear model fitted with water temperature data from USGS Gage ID: 11381500 and air temperature data from NOAA CDO Station Id: USW00024216

Paynes Creek Regression method was fitted to data from Antelope River temperature gauges

Stony Creek HEC-5Q model output at STONY CR in SSJB_SAC_Reference_062315/SAC/SAC_CL_TEMP.DSS

Thomes Creek HEC-5Q model output at THOMES CR in SSJB_SAC_Reference_062315/SAC/SAC_CL_TEMP.DSS

Upper-mid Sacramento River HEC-5Q model output at STONY CREEK in SSJB_SAC_Reference_062315/SAC/SAC_CL_TEMP.DSS

Sutter Bypass Tisdale Rotary Screw Trap montly mean water temperature 2011-2017

Bear River Regression method was fitted to data from Deer Creek temperature gauges

Feather River Regression method was fitted to data from Deer Creek temperature gauges

Yuba River Estimated mean monthly water temperature from a linear model fitted with water temperature data from USGS Gage ID: 11421000 and air temperature data from NOAA CDO Station Id: USW00024216

Lower-mid Sacramento River HEC-5Q model output at KNIGHTS LDG in SSJB_SAC_Reference_062315/SAC/SAC_CL_TEMP.DSS

Yolo Bypass Knights Landing Rotary Screw Trap montly mean water temperature 2004-2018

American River HEC-5Q model output at WILLIAM POND PARK in SSJB_AR_Reference_063015/AR/AR_CL_Temp.dss

Lower Sacramento River Measured mean monthly water temperature from USGS Gage ID: 11447650 with imputed missing values using forecast::na.interp

Calaveras River Regression method was fitted to data from Deer Creek temperature gauges

Cosumnes River Estimated mean monthly water temperature from a linear model fitted with water temperature data from USGS Gage ID: 11335000 and air temperature data from NOAA CDO Station Id: USW00023271

Mokelumne River Estimated mean monthly water temperature from a linear model fitted with water temperature data provided by EBMUD measured near Victor, CA and air temperature data from NOAA CDO Station Id: USC00045032

Merced River HEC-5Q model output at SANTA FE BR in SSJB_SJR_Reference_062915/SJR/SJR_CL_TEMP.DSS

Stanislaus River HEC-5Q model output at BLW MCHENRY BR in SSJB_SJR_Reference_062915/SJR/SJR_CL_TEMP.DSS

Tuolumne River HEC-5Q model output at GEER ROAD BR in SSJB_SJR_Reference_062915/SJR/SJR_CL_TEMP.DSS

San Joaquin River HEC-5Q model output at ABV TUOL in SSJB_SJR_Reference_062915/SJR/SJR_CL_TEMP.DSS