Migratory Corridor Water Temperature

The median proportion of days over 20°C per month

prop_temp_over_20_migr_cor

Format

dataframe with 300 rows and 4 variables

order

ordering of watersheds within salmon life cycle model

watershed

CVPIA watershed

month

integer representation of calendar month

median_p20

median proportion of temperature over 20°C

Source

Sadie Gill sgill@flowwest.com

Details

Watersheds were assigned membership to one of the following three locations to represent the migratory corridor temperature conditions. The data was summarised by calculating the median proportion of days over 20°C for each month during the available period of record.

Sacramento River, Sacramento River Below Wilkins Slough: Upper Sacramento River, Antelope Creek, Battle Creek, Bear Creek, Big Chico Creek Butte Creek, Clear Creek, Cottonwood Creek, Cow Creek, Deer Creek, Elder Creek, Mill Creek, Paynes Creek, Stony Creek, Thomes Creek, Bear River, Feather River, Yuba River American River

South Delta, Mokelumne River At San Joaquin River: Calaveras River, Cosumnes River, Mokelumne River

San Joaquin River, San Joaquin River Near Vernalis: Merced River, Stanislaus River, Tuolumne River

NA (no spawning in these regions): Upper-mid Sacramento River, Sutter Bypass, Lower-mid Sacramento River, Yolo Bypass, Lower Sacramento River, San Joaquin River